+++ src/main/java/qarenabe/qarenabe/repository/LessonRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LessonRepository extends JpaRepository<LessonRepository, Integer> {
+public interface LessonRepository extends JpaRepository<LessonRepository, Long> {
 }
