+++ src/main/java/qarenabe/qarenabe/repository/UserCourseRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserCourseRepository extends JpaRepository<UserCourseRepository, Integer> {
+public interface UserCourseRepository extends JpaRepository<UserCourseRepository, Long> {
 }
